Change incorrect WebSocket reference to Unix sockets
I believe that the reference to WebSockets is incorrect in the book. I
think that this is not actually WebSockets but instead a standard Unix
socket abstraction.

Normally, WebSockets is a protocol for establishing a two-way
communication to a backend webserver (after performing a HTTP upgrade).

In the book, only the normal request and response flow is shown.

[1] https://en.wikipedia.org/wiki/WebSocket
[2] https://en.wikipedia.org/wiki/Berkeley_sockets#Socket_API_functions

Once again, thank you for writing this book.

 I assign the copyright of this contribution to Pedro Duarte Faria.